+++ src/ZS.approx.null.np.c
@@ -41,10 +41,12 @@ static void Cintfn(double *x, int n, void *ex)
 // use R's integrate code from QUADPACK to obtain marginal likelihood
 double ZS_logmarg(double R2, int n, int d, double rscale) {
 
-  double bound=DBL_EPSILON, epsabs, epsrel, result, abserr, *work, *ex;
+  double mode = 0.0, status=1.0, root, bound=DBL_EPSILON, epsabs, epsrel, result, abserr, *work, *ex;
   int inf = 1L, neval, ier, limit=200, lenw, last, *iwork;
   SEXP Rtheta;
   C_int_struct is;
+
+
 
   if (d <= 1 || n - d <= 1) return(0.0);
 
@@ -54,31 +56,38 @@ double ZS_logmarg(double R2, int n, int d, double rscale) {
   iwork = (int *) R_alloc((size_t) limit, sizeof(int));
   work = (double *) R_alloc((size_t) lenw, sizeof(double));
 
-  PROTECT(Rtheta = allocVector(REALSXP, 4));
+  mode = find_mode_g_JZS(R2, n, d, &root, &status);  // log density at mode
+ // Rprintf("mode %lf g-mode %lf status = %lf\n", mode, root, status);
+
+  PROTECT(Rtheta = allocVector(REALSXP, 6));
   REAL(Rtheta)[0] = R2;
   REAL(Rtheta)[1] = (double) n;
   REAL(Rtheta)[2] = (double) d;
   REAL(Rtheta)[3] = (double) rscale;
+  REAL(Rtheta)[4] = (double) mode;
+  REAL(Rtheta)[5] = (double) root;
 
   ex = REAL(Rtheta);
 
   is.f = ZS_density;
   is.theta = Rtheta;
 
+  inf = 2L;  // use u = log(g) os integration range i -Inf to Inf
+
   Rdqagi(Cintfn, (void*)&is, &bound,&inf,&epsabs,&epsrel,&result,
          &abserr,&neval,&ier,&limit,&lenw,&last,iwork,work);
 
-  if (!R_FINITE(result)) {
-    Rprintf("ZS return: logBF %lf R2=%lf n= %lf d=%lf r=%lf \n", log(result), ex[0], ex[1], ex[2], ex[3]);
-  }
+//  if (!R_FINITE(result)) {
+//    Rprintf("ZS return: logBF %lf R2=%lf n= %lf d=%lf r=%lf \n", log(result), ex[0], ex[1], ex[2], ex[3]);
+//  }
 
   UNPROTECT(1);
-  return(log(result));
+  return(log(result) + mode);
 }
 
 double ZS_shrinkage(double R2, int n, int d, double rscale) {
 
-  double bound=0.0, epsabs, epsrel, result, abserr, *work, *ex;
+  double mode, logmarg, bound=0.0, epsabs, epsrel, result, abserr, *work, *ex;
   int inf = 1L, neval, ier, limit=200, lenw, last, *iwork;
   SEXP Rtheta;
   C_int_struct is;
@@ -96,7 +105,7 @@ double ZS_shrinkage(double R2, int n, int d, double rscale) {
   REAL(Rtheta)[1] = (double) n;
   REAL(Rtheta)[2] = (double) d;
   REAL(Rtheta)[3] = (double) rscale;
-
+  REAL(Rtheta)[4] = (double) mode;
   ex = REAL(Rtheta);
 
   is.f = ZS_density_shrinkage;
@@ -105,54 +114,124 @@ double ZS_shrinkage(double R2, int n, int d, double rscale) {
   Rdqagi(Cintfn, (void*)&is, &bound,&inf,&epsabs,&epsrel,&result,
          &abserr,&neval,&ier,&limit,&lenw,&last,iwork,work);
 
+  logmarg = ZS_logmarg(R2, n, d, rscale);
+  if (!R_FINITE(result) | !R_FINITE(logmarg)) result = 1.0;
+  else result = exp(log(result) - logmarg + mode);
 
-  result = exp(log(result) - ZS_logmarg(R2, n, d, rscale));
 //  Rprintf("Shrinkage return: logBF %lf R2=%lf n= %lf d=%lf r=%lf \n", result, ex[0], ex[1], ex[2], ex[3]);
   UNPROTECT(1);
   return(result);
 }
 
+// // return log(1 + exp(x)), preventing cancellation and overflow */
+// From http://www.codeproject.com/KB/recipes/avoiding_overflow.aspx
+// from  BayesFactor  bfutility
+double log1pExp(double x)
+{
+  const double LOG_DBL_EPSILON = log(DBL_EPSILON);
+  const double LOG_ONE_QUARTER = log(0.25);
+
+  if (x > -LOG_DBL_EPSILON)
+  {
+    // log(exp(x) + 1) == x to machine precision
+    return x;
+  }
+  else if (x > LOG_ONE_QUARTER)
+  {
+    return log( 1.0 + exp(x) );
+  }
+  else
+  {
+    // Prevent loss of precision that would result from adding small argument to 1.
+    return log1p( exp(x) );
+  }
+}
+
+double logExpXplusExpY( const double x, const double y )
+{
+  return x + log1pExp( y - x );
+}
+
 void ZS_density(double *x, int n, SEXP Rtheta) {
 // d is p + 1  and includes the intercept
 // prior for  1/g ~ gamma(1/2, rscale*n/2)
-   double g, R2, rscale, d, nobs;
+// x = log(g) for increased numerical stability
+   double u, R2, rscale, d, nobs, mode, shift;
    int i;
 
    PROTECT(Rtheta);
    SEXP Rex = PROTECT(duplicate(Rtheta));
 
    R2 =  REAL(Rex)[0];
-   // Hack for error with integration leading to marg that are NA - Fix
-   if (R2 > .999999)  R2 = .9999;
+
    nobs = REAL(Rex)[1];
    d = REAL(Rex)[2];
    rscale = REAL(Rex)[3];
+   mode = REAL(Rex)[4];
+   shift = REAL(Rex)[5];
+
    for (i=0; i < n; i++) {
-    g = x[i];
-    x[i]  = .5*(log(1.0 + g)*(nobs - d) - (nobs-1.0)*log(1.0 + (1.0 - R2)*g));
-    x[i] += .5*(log(.5*nobs*rscale) -3.0*log(g) - rscale* nobs/g) - lgamma(.5);
-  if (!R_FINITE(x[i]))  Rprintf("integrate: g= %lf BF= %lf R2=%lf nobs= %lf d=%lf r=%lf \n",
-            g, exp(x[i]), R2, nobs, d, rscale);
+    u = x[i] +  log(shift); // on input  log(g) = u; shift by mode
+    x[i] = -mode;   //  subtract mode of f on log scale
+    x[i] += .5*(log1pExp(u)*(nobs - d) -
+               (nobs-1.0)*log1pExp(u + log(1.0 - R2)));
+    x[i] += .5*log(.5*nobs*rscale) -lgamma(.5) -
+            (1.0 + .5)*u - .5*rscale*nobs*exp(-u) ;
+    x[i] += u;   // log jacobian
+//  if (!R_FINITE(x[i]))  Rprintf("integrate: u= %lf logBF= %lf R2=%lf nobs= %lf d=%lf r=%lf mode=%lf shift%lf\n",
+//            u, x[i], R2, nobs, d, rscale, mode, shift);
 
     x[i] = exp(x[i]);
   }
  UNPROTECT(2);
 }
 
+void ZS_density_g(double *x, int n, SEXP Rtheta) {
+  // d is p + 1  and includes the intercept
+  // prior for  1/g ~ gamma(1/2, rscale*n/2)
+  double g, R2, rscale, d, nobs, mode;
+  int i;
+
+  PROTECT(Rtheta);
+  SEXP Rex = PROTECT(duplicate(Rtheta));
+
+  R2 =  REAL(Rex)[0];
+
+  nobs = REAL(Rex)[1];
+  d = REAL(Rex)[2];
+  rscale = REAL(Rex)[3];
+  mode = REAL(Rex)[4];
+
+  for (i=0; i < n; i++) {
+    g = x[i]; // on input
+    x[i] = -mode;   // "divide" by mode or subtract on log scale
+    x[i] += .5*(log(1.0 + g)*(nobs - d) - (nobs-1.0)*log(1.0 + (1.0 - R2)*g));
+    x[i] += .5*(log(.5*nobs*rscale) -3.0*log(g) - rscale* nobs/g) - lgamma(.5);
+    if (!R_FINITE(x[i]))  Rprintf("integrate: g= %lf BF= %lf R2=%lf nobs= %lf d=%lf r=%lf mode=%lf\n",
+        g, exp(x[i]), R2, nobs, d, rscale, mode);
+
+    x[i] = exp(x[i]);
+  }
+  UNPROTECT(2);
+}
+
 void ZS_density_shrinkage(double *x, int n, SEXP Rex) {
   // d is p + 1  and includes the intercept
   // prior for  1/g ~ gamma(1/2, rscale*n/2)
-  double g, R2, rscale, d, nobs;
+  double g, R2, rscale, d, nobs, mode;
   int i;
 
   PROTECT(Rex);
 
-  R2 =  REAL(Rex)[0];
-  nobs = REAL(Rex)[1];
-  d = REAL(Rex)[2];
+  R2     =  REAL(Rex)[0];
+  nobs   = REAL(Rex)[1];
+  d      = REAL(Rex)[2];
   rscale = REAL(Rex)[3];
+  mode   = REAL(Rex)[4];
+
   for (i=0; i < n; i++) {
-    g = x[i];
+    g = x[i];  // on input
+    x[i]  = -mode;
     x[i]  = .5*(log(1.0 + g)* (nobs - d) - (nobs-1.0)*log(1.0 + (1.0 - R2)*g));
     x[i] += .5*(log(.5*nobs*rscale) -3.0*log(g) - rscale* nobs/g) - lgamma(.5);
     x[i] = exp(x[i])*g/(1.0+g);
@@ -209,6 +288,35 @@ double LogBF_ZS_null(double R2, int n, int d) {
     }
   return(NA_REAL);
 }
+
+double find_mode_g_JZS(double R2, int n, int d, double *root, double *status) {
+  int k;
+  double aux, a, b, c, mode;
+// d is dimension of model with intercept
+
+void posroot(double a, double b, double c, double *root, double *status);
+double lik_null(double g, double R2,int n,int k);
+
+k = d - 1;
+aux=-(1.0-R2)*((double)k+3.);
+a=(double)n-4.-(double)k-2.*(1.0-R2);
+b=(double)n*(2.-R2)-3.;
+c=(double)n;
+
+a=a/aux;
+b=b/aux;
+c=c/aux;
+
+posroot(a,b,c,root,status);
+
+
+if (*status != 1.0)  *root = n/20.0;  //should not occur but fix
+
+mode = lik_null(*root,R2,n,k);  // this is log marginal at mode
+
+return(mode);
+}
+
 
 double lik_null(double g, double R2, int n, int k){
 /* this computes log(likelihood  x prior), where the likelihood is marginal