+++ src/bas.h
@@ -144,6 +144,7 @@ double ZS_logmarg(double R2, int n, int d, double rscale);
 double ZS_shrinkage(double R2, int n, int d, double rscale);
 void ZS_density(double *x, int n, SEXP Rtheta);
 void ZS_density_shrinkage(double *x, int n, SEXP Rtheta);
+double find_mode_g_JZS(double R2, int n, int d, double *root, double *status);
 double E_ZS_approx_null(double R2, int n, int d);
 double BIC(double Rsquare, int n,  int p, double SSY);
 double AIC(double Rsquare, int n,  int p, double SSY);