+++ NEWS.md
@@ -1,8 +1,10 @@
-#  BAS 1.4.7 June 23, 2017
+#  BAS 1.4.7 July 30, 2017
 
 ## Updates
 
-* Added new method for `bas.lm` to obtain marginal likelihoods with the Zellner-Siow Priors for method "ZS-null" using QUADMATH routines for numerical integration.  The optional hyperparameter alpha may now be used to adjust the scaling of the ZS prior where g ~ G(1/2, alpha*n/2) with a default of alpha=1 corresponding to the ZS prior used in Liang et al (2008).
+* Added new method for `bas.lm` to obtain marginal likelihoods with the Zellner-Siow Priors for "prior= 'JZS' using QUADMATH routines for numerical integration.  The optional hyperparameter alpha may now be used to adjust the scaling of the ZS prior where g ~ G(1/2, alpha*n/2) as in the `BayesFactor` package of Morey, with a default of alpha=1 corresponding to the ZS prior used in Liang et al (2008).  This also uses more stable evaluations of log(1 + x) to prevent underflow/overflow.
+
+* Priors `ZS-null` and `ZS-full` for bas.lm are planned to be deprecated with numerical integration providing more accurate answers.  
 
 * replaced math functions to use portable C code from Rmath and consolidated header files